move and minor refactor PrismaAdapter
