+++ packages/jobs/src/adapters/PrismaAdapter/__tests__/PrismaAdapter.test.ts
@@ -1,10 +1,8 @@
 import type { PrismaClient } from '@prisma/client'
 import { describe, expect, vi, it, beforeEach, afterEach } from 'vitest'
 
-import type CliHelpers from '@redwoodjs/cli-helpers'
-
-import { DEFAULT_MODEL_NAME } from '../../consts'
-import * as errors from '../../errors'
+import { DEFAULT_MODEL_NAME } from '../../../consts'
+import * as errors from '../../../errors'
 import { PrismaAdapter } from '../PrismaAdapter'
 
 vi.useFakeTimers().setSystemTime(new Date('2024-01-01'))
